viewからcontrollerに移しました。
