+++ app/src/main/java/com/leverages/careerticketcafe/controllers/LoginActivity.kt
@@ -1,13 +1,13 @@
-package com.leverages.careerticketcafe.views
+package com.leverages.careerticketcafe.controllers
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.leverages.careerticketcafe.R
 
-class SeminarListActivity : AppCompatActivity() {
+class LoginActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_seminar_list)
+        setContentView(R.layout.activity_login)
     }
 }
