+++ app/src/main/java/com/leverages/careerticketcafe/controllers/SeminarDetailActivity.kt
@@ -1,13 +1,13 @@
-package com.leverages.careerticketcafe.views
+package com.leverages.careerticketcafe.controllers
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.leverages.careerticketcafe.R
 
-class LoginActivity : AppCompatActivity() {
+class SeminarDetailActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
+        setContentView(R.layout.activity_seminar_detail)
     }
 }
