+++ app/src/main/java/com/leverages/careerticketcafe/controllers/SeminarListActivity.kt
@@ -1,13 +1,13 @@
-package com.leverages.careerticketcafe.views
+package com.leverages.careerticketcafe.controllers
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.leverages.careerticketcafe.R
 
-class TopActivity : AppCompatActivity() {
+class SeminarListActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_top)
+        setContentView(R.layout.activity_seminar_list)
     }
 }
