+++ app/src/main/java/com/leverages/careerticketcafe/controllers/TopActivity.kt
@@ -1,13 +1,13 @@
-package com.leverages.careerticketcafe.views
+package com.leverages.careerticketcafe.controllers
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.leverages.careerticketcafe.R
 
-class NewsDetailActivity : AppCompatActivity() {
+class TopActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_news_detail)
+        setContentView(R.layout.activity_top)
     }
 }
