+++ app/src/main/res/layout/activity_menu_detail.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.MenuDetailActivity">
+    tools:context=".controllers.MenuDetailActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
