+++ app/src/main/res/layout/activity_news_detail.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.NewsDetailActivity">
+    tools:context=".controllers.NewsDetailActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
