+++ app/src/main/res/layout/activity_news_list.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.NewsListActivity">
+    tools:context=".controllers.NewsListActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
