+++ app/src/main/res/layout/activity_seminar_detail.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.SeminarDetailActivity">
+    tools:context=".controllers.SeminarDetailActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
